Replace CHECKSIZE with getData() checks
TiffEntry parsing can lead to out-of-bounds access. Guard
against that with getData() checks instead of CHEKSIZE and
bail out of processing IFDs or Entries when the checks
fail but don't bail out of the whole file because of that.
Raw formats being what they are they are still parseable
even if some entries are broken.