TIFF: avoid inifinite looping SubIFDs in TIFF
At the root level TIFF has the notion of the next IFD
and if that's made into a loop we get into a loop of
heap-allocating calls, eventually running out of memory.
Fix that just by limiting the max number of SubIFDs the
root one can have to 100 which should be more than plenty
for any sane file.