fix crash in DNG opcode
OpcodeFixBadPixelsList would read the number of points
and rectangles as a signed int and then allow negative
numbers. This would eventually be used in a bounds check
allowing a large negative number to offset a large
positive one and then cause an out of bounds access.
Just reject negative counts.