avoid out of bounds with TIFF_UNDEFINED
TIFF_UNDEFINED is 8bit untyped data that we then allow
to be accessed with getShortArray() and getIntArray().
Instead of allowing count shorts or ints only access
count/2 shorts and count/4 ints. The caller is still
required to make sure he doesn't access more than
count/2 shorts or count/4 ints in the returned array.