NIKON D300: all the 12bit/14bit modes.
Reported by Jack Massey on IRC, with a sample.
The sample looked very dark, but had purple highlights.

We should probably just stop loading NEF's unless that specific
mode is listed in cameras.xml, else we are quite likely to
be providing broken images. Especially since that is what we do
for the cameras that are entirely not listed in cameras.xml