NefDecoder: stop providing the rope for users to hang themselves.
This is basically the same problem as with Rw2Decoder.

It looks like all the nikon cameras has different modes -
14-bit compressed, 12-bit compressed,
14-bit uncompressed, 12-bit uncompressed.

And the problem here is that 14-bit and 12-bit have different white levels.

Since NefDecoder::checkSupportInternal() just silently fallbacks
to the generic entry, if present (with mode=""),
this has result in silently-garbled images.

Let's not do that.