nikon modes inventory check :(
For all the nikon cameras, if only the entry for 14-bit mode
are present, this *will* finally break support for 12-bit modes.
(as in, rawspeed will stop silently producing subtly broken images)

Same for 12-bit modes.

This will probably break following cameras:
(these have white="0", so i can't guess what they are)
        <Camera make="NIKON CORPORATION" model="NIKON 1 J1" decoder_version="4">
        <Camera make="NIKON CORPORATION" model="NIKON 1 J2" decoder_version="4">
        <Camera make="NIKON CORPORATION" model="NIKON 1 S1" decoder_version="4">
        <Camera make="NIKON CORPORATION" model="NIKON 1 V1" decoder_version="4">
(these are apparently actually 16bit, but report as (?)-bit)
        <Camera make="NIKON" model="COOLPIX P330" decoder_version="5">
        <Camera make="NIKON" model="COOLPIX P7700" decoder_version="5">
        <Camera make="NIKON" model="COOLPIX P7800" decoder_version="5">

The basic idea for this commit:
$ grep -ri "<camera" cameras.xml | grep -i nikon | grep -iv "mode=" | sort
For each camera in that list ^ look for the presence of same entry
but with modes, and if any - add missing modes.

If the white level is 14-bit integer - the entry is for mode="14bit-*"
And it is safe to assume that the entry for 12bit is missing

If the white level is 12-bit integer - the entry is for mode="12bit-*"

In the end, there should be no <Camera> entries for nikon cameras with
missing / empty mode tag.