HuffmanTable: rewrite Huffman decoder related code
This is pretty much a rewrite of all Huffman decoding related code.

There have been several copies of this functionality with minor
differences (and different bugs). They are all replaced by one new
implementation in HuffmanTable.h. Code documentation of the inner
workings has been significantly increased.

Great emphasis has been put on optimizing the hot paths of the decoding
on down to the CPU instruction level. They are responsible for somehere
nearly the complete decoding time of a standard LJPEG based raw. There
is an optimized, unified lookup mechanism, replacing the old 'numbit'
and 'bigTable' funtionality.

To give an example: the number of executed instructions and executed
branches went down by about 30% - the runtime by 10%-20% for some sample
CR2.

This cleanup is not quite complete. The LJPEGDecoder and friends clases
using this code have only been changed minally to support the new code.
This will follow in a separte commit.