LJpegDecompressor cleanup
Replace 6 similar versions of the guts of the  LJpegPlain::decodeScan()
function (2 of which were unused) with 1 generic version that uses
integer template parameters to generate 4 specialized versions that run
just as fast as the "havily optimized" ones before did. This drops the
total line count by about 800.

LJpegDecompressor has now exactly 1 (pure) virtual function that gets
overriden in the two remaining subclasses LJpegPlain and
HasselbladDecompressor. The copy of parseSOF in HasselbladDecompressor
has been removed. All BitPumps are now created on the stack and exactly
where they are used. 'startDecoder' has been renamed to 'decode()' to
better describe that this is not starting a thread or something like
that.

This LJpegDecompressor cleanup is still a work in progress.