Cr2Decoder/LJpegPlain simplification, documentation and cleanup
1. Turn the main decoing loop inside-out:
Instead of going over the raw image "virtually" line-by-line and deal
with the calculation of the current destination position, the current
predictor and the current slice from the inside, do it in the more
natural way: slice after slice as decoded from the jpeg stream. This
allows to have 3 fewer variables to track where we are in the decoding
process (less state to reason about).

2. Simplify the handling and improve documentation of the different
Canon quirks:
* 80D mraw frame vs image size mismatch (now autodetected)
* double-width/half-height 5Ds frame layout
* 6D flipped width/height
* remove the unnecessary (and buggy) skipY/skipX logic

This removes about 50 lines of code to reason about.