Remove the usage of invalid CFA_GREEN2
This fixes #34. The actual number '3' is still kept as a dummy tag to
keep the ColorFilterArrayTest.ToDcrawAndBack unit test happy.

The string output of this dummy tag is now "UNKNOWN" therefore the
string length check unit test had to be slightly adjusted.

This change 'breaks' 3 files in rstest by actually fixing them: the old
"GREEN2" color is now reported as "GREEN".