Introduce AbstractTiffDecoder
AbstractTiffDecoder is a new base class of all RawDecoders that had a
TiffIFD* mRootIFD. This new base class deals with the memory ownership
task for this TiffIFD. It uses a TiffRootIFDOwner, wich is just a
unique_ptr to handle that. This completes the memory management cleanup
of the TiffIFDs started in one of my first prs. This also properly fixes
any possibly still remaning leaks of that memory in case of an
exception.

It allows to remove lots of now empty con-/de-structors in the affected
derived classes. And it fixes a couple of situations where the derived
class forgot to overrwide getRootIFD() from the RawDecoder class, which
is exactly the type of problem that comes from the code douplication I'm
getting rid of here.