Consolidate more common code of the Tiff based decoders
Looking into the discussed getTrimmedString() change, I noticed _a lot_
of duplicated code extracting the MAKE and MODEL strings from the tiff.
So all sub classes of the new AbstractTiffDecoder do have
(unsurprisingly) a lot more in common than only sharing the mRootIFD
member. All used the same 7-line code pattern to get to the MAKE/MODEL
strings (twice!). And almost all implementations of
checkSupportedInternal() were completely identical.

This patch provides a default implementation for that function in the
base class. It further introduces a little helper called getID() to the
TiffRootIFD class that returns a make/model string pair. These strings
are automatically trimmed. This removes almost all former uses of
getString() and also some of the TrimSpaces() calls. So the default
trimming is moved from all getString() calls to only the ones dealing
with the default access to MAKE/MODEL strings. This makes the discussion
about introducing getTrimmedString obsolete.

Another 200 lines of duplicated code gone.

Also minor improvement of the MOSDecoder that also gets a
TiffRootIFDOwner to get rid of the need for a delete in its destructor.