RawImageDataU16: fix data acess inside fixBadPixel
The code was designed to access the complete image but then used
getData() instead of getDataUncropped(). This worked for all files up
until now because the 'redefinition' of what getData() means usually
happened after fixBadPixel() was called. Not with the DNG code, thogh.

Note: this 'redefinition' of the meaning of getData() at some 'random'
point during the decoding process is a really bad interface design.