Coverity: fix SIGN_EXTENTION with clampBits
CID 1402217

Improve clampBits interface by fitting it more to the problem at hand:
This function is solely used to clamp some int into an unsigned short
(data type of 16 bit raw image). So have it return a ushort16 and add an
assert(n <= 16). This would effectively allow find a bug if someone
would accidentally clip to 17 bits and then assign the resulting 32-bit
int to a 16bit one.

Fixes #84