Remove unnecessary eplicit default con/de-structors and operators
Explicitly specifying a default constructor only makes sense in cases
where the compiler does not do so anyway (e.g. where a non-default
constructor is present and the default one is also required). If not
needed, they just add noise to the code base. As long as the code
compiles, removing them has zero influence on the semantic of the code.