Fix BlackAreaTest unused variable warning (broken clang compile)
The removal of the explicit default destructor of the trivial BlackArea
struct caused clang to complain about an unused variable in the test.

This removes all tests from the BlackAreaTest.cpp that merely test if
the compiler is able to produce proper default asigment operators and
default assigment constructors. The compilers have their own regression
tests for that. There is only one line of code in this class and that is
the asignemnt of 3 variables, checking whether that assignment worked is
the only thing there is to test.