IiqDecoder: implement slice size computations.
Offsets are not guaranteed to be in monotonically increasing order.
So for each element of 'offsets', we need to find a element
(which is located after the current element), which specifies
larger offset. And only then by subtracting those two offsets
we get the slice size.

This seems to be the right thing to do (C)
At least it is valid for all the current samples.

Unless they intentionally encode something that is supposed
to be present in more than one slice
by just specifying one slice [offset] twice...

Is there really no such info in the entries?
Tag 0x223 does not seem to be it.