Skeleton of Phase One Sensor Corrections
dcraw has a function called phase_one_correct, which applies some
corrections to the raw data based on metadata embedded in the RAW
file.  I would like to implement any of these that I can find RAW
files serving as examples for, as they can make a significant
difference on the output (the quadrant multipliers, in particular, can
cause visible seams to appear in the rendered image if not correctly
applied).  This commit just adds some code to read the extended
metadata, with the skeleton of a function for one particular type of
correction and room to add more.