HasselbladDecompressor::decode(): an attempt to sanitize pixelBaseOffset.
The math *suggests* that values out of that range will result
in predicted pixel being unfitting for ushort16, i.e. 0..65535
Note however that check for predictor value is not added,
because all the current Hasselblad 3FR samples break.

In particular, i plugged this into wolframalpha:
  32768 + x +  32768 >= 0, 32768 + x +  32768 <= 65535,
  32768 + y + -32768 >= 0, 32768 + y + -32768 <= 65535
and used the minimal value the x can take, and the maximal
value y can take as the limits.

Assumption: -32768..32768 is the range of values produced
by HasselbladDecompressor::getBits().