FujiDecompressor::FujiHeader: sanitize blocks_in_row wrt. raw_width and block_size
Fuji compressed RAF's are split into strips (blocks):
       0
     0 /----|----|---#\
       | s0 | s1 | s2#|
height \----|----|---#/
                    /|\
                  width

Each strip spans entire height, but only part of the width.
For padding purposes, all the strips encode the same
amount of columns, but the final image may be smaller
than that, so some columns of the last slice may be ignored.

There is an assertion in FujiDecompressor.h:91:
	assert(h.block_size * h.blocks_in_row >= h.raw_width);
It ensures that last strip does actually contain enough pixels
to finish the row. (Well, it checks that all the slices produce
at least enough columns for the entire image).

But the FujiHeader sanitization did not actually sanitize
that case, so it was possible to hit the assert.