NefDecoder::decodeMetaDataInternal(): harden serial number parsing.
Two things: if we use signed char, then if we get [-128..-1],
after modulo we will still have the negative number, which
will can cause drastic overflow even if the length is just 2.
I guess in all normal cases each symbol will either be '0'..'9'
or 0..9, so we should be good even with unsigned.

And, since we use fixed-size uint32, there is some upper length
of the serial number string we can convert. And it is 8.
Given that it seems they are 7 in real samples, we are good.