KdcDecoder: read sensor dimensions from the appropriate entries.
These two entries, for two[*] samples that are currently working,
result in the very same final image dimensions that we'd get
after incrementing. But avoiding increment avoids overflow.

* Kodak P880 and Z1015 IS
  Other raws that should be handled by this path are mis-decoded already.

Z1015 IS hash changes though, one last row is gone.
Maybe we are still looking at the wrong tags.
But i'm not seeing the obvious visual symptom (stairstep)
of a wrong width/pitch, so not sure yet.