Cr2Decompressor::decodeN_X_Y(): handle yet another unsufficient image size
If we don't have enough slices, we won't fill the entire image area
we were supposed to, so those pixels will be uninit garbage.

I really hate this code. It might be the ugliest in the whole library.
Other than the Fuji decompressor. This really needs to be rewritten,
without all these branches nested deep within loops.