DngDecoder::handleMetadata(): 'fix' whitelevel handling.
And we have a first real bug here.

Whatever we did with that whitelevel in the case of
floating point image is just wrong.
White level should be 1.0f there, while we ended up with ~0U.
But we can't actually represent the right value of 1.0f...

So i think pretending to be a 16-bit image is the least wrong thing to do.
In practice, there aren't any samples where this value
wasn't overwritten with the actual whitelevel from metadata.

../src/librawspeed/decoders/DngDecoder.cpp:510:22: runtime error: implicit conversion from type 'unsigned long' of value 4294967295 (64-bit, unsigned) to type 'int' changed the value to -1 (32-bit, signed)
    #0 0x2bdc36 in rawspeed::DngDecoder::handleMetadata(rawspeed::TiffIFD const*) /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decoders/DngDecoder.cpp:510:22