VC5Decompressor: sink MAX_NUM_CHANNELS into VC5Decompressor.
While i understand that VC5_TAG_ChannelCount specifies
the actual channel number, there is not a single sample
where it isn't 4, so we don't actually know that
we handle any other channel count correctly.
Should the sample appear, this can be relaxed.

Also, don't forget to sanitize 'iChannel' when reading.