VC5Decompressor::Wavelet::reconstructLowband(): vertical, middle rows: move common 'pass'es together.
Associativity, basically:
a + ((b) >> 3) == (((a << 3) + b) >> 3)
Though, a can (is in one example) be negative,
so we can't shift, and need to multiply.

This is rather questionable i must admit.
Either that was a "premature" optimization to avoid this
associative operation, or that was *really* intentional,
The difference is important, because these variants
are *different* as per C++.
https://godbolt.org/z/a9YxmG
https://rise4fun.com/Alive/0Ml