BitStream<>::fillSafe(): adjust size of accepted past-the-end pos overrun.
In some cases, the input bits end up being tightly packed into
the input buffer, with no spare padding 'unused' bits at the end.
Most prominently that happens for some Nikon '.NEF' raws.
A special care needs to be applied to be still able to extract
those last few *real* bits (as opposed to erroneously marking
the raw as corrupted/truncated), and ideally with no perf impact.

Here, the solution is to *pretend* that the input buffer is
slightly larger than it is, and pretend that the 'padding'
is all-zeros. Thought indeed, if those empty 'padding' bits
will be read/used, the decoding may produce something wrong.
(i *thought* i saw that in
https://github.com/darktable-org/rawspeed/issues/156
but i can no longer reproduce that. cosmic ray bitflips?)

*How* slightly larger is a question. Roughly, enough so that
the last few(*) bits can still be accessed, while pretending
that the `fillLevel` is `32`.
(*) This works for 'last few' being ~20 bits, but i'm not sure
    this won't need tuning for smaller definition of 'last few'.

The alternative approach could be to track the remaining
number of last few *actual* bits, and keep pretending that
the buffer did not finish until all those bits were consumed.
That would still provide all the same termination guarantees,
but the code would be really ugly.
Or maybe that should be done for the DEBUG build type.

Fixes https://github.com/darktable-org/rawspeed/issues/128
See https://redmine.darktable.org/issues/12208