SonyArw2Decompressor::decompressRow(): _imax and _imin can't be equal.
We have 98 bits, and we need to decode 16 pixels, with 7/bit pixel.
But we know the values of the minimal and maximal pixel,
we use those values, and voila, we have 16 pixels!

BUT, if the same pixel is marked as being minimal and maximal,
then only that one pixel will use that extreme value.
So we are still one pixel short. That appears to be an invariant.

And now, decompress() is no longer non-throwing.