NefDecoder: rewrite NEFIsUncompressed() to accept some small padding
Happens in Nikon Z7 / Z6 uncompressed raws.
The actual size of the padding to be accepted is a guess.
I'm unable to come up with any better logic other than this.

Note that this is all unsafe (as in can have false-positives)
for the very same reason why this limit for padding exists:
some *compressed* NEF's also pass this check.