Split naive implementation of rawspeed_get_number_of_processor_cores() into a helper library.
That function has to be defined NOT by the rawspeed library itself
so that the users have a choice of implementation.
But at the same time that helper library needs to be linked
after the rawspeed library. AND it needs to be 'linked'
against the rawspeed library for include paths.
Thus keeping it as a small helper library avoids cyclic deps.