CMake: don't FORCE-set the RAWSPEED_STANDALONE_BUILD.
The main reason why that flag is there is so that we don't
retain the RAWSPEED_SOURCE_DIR in sources, which would be
bad for reproducibility reasons (if the upper project is installed),
and those paths would likely be never found after install anyway.

But there is an upcoming usecase of being built as subproject,
but not installed, used as-is. And for that RAWSPEED_SOURCE_DIR
is ok. Thus let's allow to override our detection here.