LNT: fix collection of compiler/linker time, compiler stats.
We *have to* collect these compile/link time, compiler stats globally.
The modules that process captured stats only look for stats
in the directory of the test executable, and thus they are extremely
optimistic as in e.g. they ignore compile time of the librawspeed,
since there are no test targets in that dir.
And they are pessimistic in the directory with multiple test targets,
since every one of them will have the same values.
So let's just capture and aggregate them for the entirety of the RawSpeed.