SonyArw1Decompressor: for a single pixel we will at most consume 32 bits
For getDiff() to consume maximal number of bits, we must hit the
'while (len < 17 && !bits.getBitsNoFill(1))' loop, and for that
the len must be 4 before it, which requires that first two bits
we read are 0b00. Thus we at most consume 15 bits within decompress()
and 17 in getDiff(), i.e. 32 in total.

https://godbolt.org/z/AQ4NFt