SamsungV1Decompressor::decompress(): don't report which pixel is bad
We just kinda don't care. All bets are off on garbage input.
One thing why it may matter is since we no longer require
to output the actual row/col, maybe the compiler will be smart
enough to switch to pointer iteration.
Here this does not really measure to anything..