Kill BUFFER_PADDING.
A configuration with BUFFER_PADDING!=0 is not even build-tested,
let alone fuzz-tested. While i'm very sure that not having to
ensure that you can read 8 bytes past the end of given buffer
gets rid of some branching, it also is not really correct thing
to do because that will blow up on first malformed raw,
and will surely explode under fuzzer.
Also, it prohibits usage of any other buffer provider other than Buffer.

It is not okay to trade correctness for performance.
There are other ways to gain that perf win.