NefDecoder: rewrite readCoolpixSplitRaw()
Initially i just tried to migrate it to getU16DataAsUncroppedArray2DRef(),
but it was such a mess it was cleaner to
just completely rewrite it while there.

The even-odd input line layout is kinda ugly,
but with a bit of thought we can still serialize
per-line outputting.

And due to the same basic idea that is applied to
precompute the position of first odd row
in future it will be possible to parallelize
all such uncompressed decompressors.