Cr2Decompressor::decodeN_X_Y(): rewrite debranching to be actually correct
That was too simple to be true in full generality, even though it works
on all the existing samples. There was no logic to ensure that that
innermost loop actually runs, so that could deadlock. Though i'm not
sure this actually fixes the bigger problem, we'll see what fuzzer says.