SamsungV2Decompressor: don't template on OptFlags
This results in major bloat (x16) in terms of the code that needs
to be compiled (and fatally decreases code coverage),
but on whole 3 RPU affected samples, produces no measurable benefit
(<1%, below noise level). So let's not do that.

This partially reverts commit c514780e40c5cf481515aefdfbb3146fbb3935a4.