Syncronize fixBadPixel() in RawImageDataFloat with RawImageDataU16 variant
I think there are pricipally three problems

a) These images can be either CFA based or not so we should use the step as
   in RawImageDataU16

b) I may not be right but i think also in this code we must use getDataUncropped
   as the coordinates are in the active area and not in the default crop.

c) The total_pixel must not use the dist but the weight.

After fixing these, RawImageDataFloat::fixBadPixel() basically matches
the RawImageDataU16::fixBadPixel(), which must mean the fixes are correct.