HuffmanTableLUT: diff_l==16 still may fit fully decoded into LUT
We were erroneously/accidentally giving up on it.

Unlike all other diff_l's, there can be only one difference value: -32768,
and as it can be seen in AbstractHuffmanTable::processSymbol(),
if diff_l==16, we don't even read diff value from bit stream,
but maybe skip it's length.

So TLDR is, for the purpose of LUT, diff_l==16 can be encoded
as long as code_l <= lookupDepth.

In RPU test-suite this only is only triggered by Nikon/D80/DSC_0143.dng,
and the perf impact is within the noise. But still.