AbstractDngDecompressor: uncompressed: don't apply big-endian override to floating-point data
I.e. if we've got 24bit fp, instead of always reading it as big-endian,
even if the DNG is little-endian, we will now respect the DNG endianness.

It isn't obvious if that's fully correct, but at least on the image at hand
it is more compatible with DNG SDK.

Fixes #252.