Require an aligned allocation function to be present, prefer `aligned_alloc()`, drop `_mm_malloc()`
On linux we should always have `aligned_alloc()`,
on macOS `posix_memalign` seems to be always present,
and on windows `_aligned_malloc()` seems to be always present,
so this shouldn't break anywhere?

While it would be nice to use `std::aligned_alloc()`,
at least on OSX that will require macOS 10.15.