VC5Decompressor: HighPassBand::decode(): refactor, -3%
Instead of having a single flat loop that assumes that the output data
has no padding between rows, which is true now, but not a feature,
do a 2D loop over each output pixel, and fill it.
IMO this is a better structure anyways.

This measures as -3% perf improvement both in single-thread
and and 32-thread usage..