VC5Decompressor: combineFinalLowpassBands(): don't collapse loops
In any realistic situation the number of rows (usually at least ~3k)
will greatly outweigh the number of threads (8..128?),
so we don't really win anything by not operating on the whole
rows each time, but by not operating on the whole rows (or at least,
not starting from the 0'th column) each time means that we can't really
predict from which column we start, and thus how aligned pointers are.

This seems to measure as performance-neutral.