NORangesSet: rewrite for performance
I didn't measure in this particular case (famous last words!),
but it really does seem like linear scan over the entire set
isn't the most optimal approach.

Instead, try to partition the set, since it is sorted,
and check that there is no overlap with elements
at the partition point, and the element before that.

This is functionally equivalent to the naive implementation.