[CMake] When building for APPLE, and not targeting 11.3+, don't even look for `aligned_alloc()`
See https://github.com/darktable-org/darktable/issues/10087

While generally `CHECK_CXX_SYMBOL_EXISTS()` would work here,
when backdeploying on OSX the symbol is there on the build host,
but possibly not on the deploy target. There's a clang diagnostic
that makes it work, but only when `-w` was not passed.

I'm not seeing any reasonable way to workaround `-w`,
so let's just hardcode it.