Import color matrices from darktable's `adobe_coeff.c` (Fixes #327)
This is a(n incomplete) Minimal Viable Product.
For all intents and purposes, unless specified otherwise,
it's intended to be identical to the `ColorMatrix2` DNG tag.

Many more things (e.g. vendor/illuminant/denominator attributes)
could be added once needed, but this is the minimal thing that just
completely subsumes the entirety of darktable's `adobe_coeff.c`

The values are implicitly fixed point with denominator of 10'000,
this is done intentionally (floating point is just nasty),
and matches the contents (or rather, the type) of DNG's `ColorMatrix2` tag
(which is `SRATIONAL`).

Illuminant is implicitly `21 (D65)`.
Vendor is mostly `"Adobe"`.

The layout also matches DNG's `ColorMatrix2` tag - row per plane,
in-row columns are in order.

The port was automatically performed via: https://gist.github.com/LebedevRI/6dcf7810273ac20e53d851361ff81cc3
The results were checked via: https://gist.github.com/LebedevRI/4910df853eef0d22947f0a38eeb097fe

Not all cameras for which there is an entry in `adobe_coeff.c`
have an entry in `cameras.xml`, those matrices are not ported.

This adds rudimentary support for reading color matrices out of DNG's,
but only if the illuminant is 21, and the srational has divisor of 10'000.
Both, especially the second one, restrictions are bugs,
but they should be good enough for now.