BitStreamReplenisherBase: require that there is at least one quantum in input
We will load up to MaxProcessBytes bytes from the input each time
we replenish the buffer. While at the end we might genuinly run out
of the input, we should have plenty the first time we replenish.

Notably, this is not exactly true for JPEG, but in reality
the lower bound is so small that i don't think this will ever come back
and bite us.