Cr2Decompressor: Cr2OutputTileIterator: just maintain `outPos` state
The original arithmetic is nerdy and hard to grasp, but divisions
are slow, and more importantly, it is based on several more preconditions
that we do not check, e.g. it requires that the final CR2 slice
produces at most one vertical strip, or at least it is not wider than
the first CR2 slices. Let's just maintain state.

This is performance-neutral.