Undo some of the recent pointer-to-byte changes in ASan/MSan wrappers
They are meant to be thin wrappers for the C ifaces that
take pointer-to-void, and does not really care about the pointed-to data,
it never even dereferences the pointers in the traditional sense.

We don't really win anything by being more strict here,
but we end with a lot more of the ugly `reinterpret_cast<>()`.