Refactor Huffman Table fuzzers
It is not useful to have a dedicated fuzzer per each bit pump,
the pumps themselves aren't particularly noteworthy,
we might as well just decide which one we'll use based on the input.

While the mode (full decode or not) is interesting,
it's a property of the huffman table flavor,
much like the 16-bit ljpeg fix, and more importantly,
i'm hoping to add a few more flavors.

So having dedicated fuzzers for each is just not sustainable.