HuffmanTableLookup: `maxCodeOL` does not need to be wider than the code
This bit of logic is very finicky,

The question is, how do we write the workloop
in `finishReadingPartialSymbol()`, so that said loop is "best"/performant
with key problem being, the code can genuinely be '0b0 @ len 1`,
and if we have no codes of e.g. len 1,
and thus the maximal code at len 1 is, what? 0?

So, even though we know when to stop reading bits
(due to the codes being monothonically increasing),
how do we gracefully deal with such gaps?

The existing implementation was using a double-width types
for the `maxCodeOL`, which was fine for 16-bit codes,
but will be more questionable for 32-bit codes.
In existing implementation, no code could ever be `0xFFFFFFFF`,
because that is a 32-bit value, and codes are 16-bit.

But, we have an invariant that can help us: all-ones code
can only happen at the largest code width, but not in the middle,
so even if the current code is smaller than the max-code at that len,
and normally we'd stop reading, we should still continue reading.

So we don't need the double-width here. Yay.
Fuzz testing seems to agree, but we'll see.