TableLookUp: improve handling of non-monotonic LUT's
Some LUT's are non-monotonic, but are just N LUT's concatenated together.
This is the case in e.g.:
https://discuss.pixls.us/t/strange-pixels-from-rawsie-dng/36485

It's trivial to detect when we "transition" from one LUT to another,
since the value suddenly decreases. It makes sense to at least
not "blur the edges" between them, by treating such cross-over points
just like we'd treat the real hard edges of the LUT,
by reusing the center value instead of the neighbours values.

That being said, i'm not fully sure this is the proper fix,
because there are still few visual glitches left with
on the test image.