Allow templating PrefixCodeLUTDecoder's underlying backing decoder
LUT decoder is two-tiered, it acts as a fact "caching" front-end,
but not always succeeds, in which cases it fallbacks
to another implementation, currently to the Lookup decoder.

But while the LUT decoder is fine with arbitrary prefix codes,
Lookup decoder specifically requires Huffman codes,
which is not great, so it would be great to be able to
replace it with something else when needed.