Slap buffering ontop of MD5 hasher interface (-1%)
Brute-ness of the simple `md5_hash()` iface came up in e.g.
e7e245cdb65888ed805c51f5eb6ba1cd791045cf - we might want to
hash data that is generally sequential, but has some padding
between some parts of it, and that's not really possible
with the old iface.

Humorously, the new code is faster just a tiny bit.