LJpegDecompressor::decodeN(): actually skip restart marker, restart intervals are supported
And thats the magic - the actual byte stream for all of the
MCU's of the restart interval is followed by a JPEG Marker
with the restart interval's counter (modulo 8)
(except for the last restart interval), and we need to skip it
to get to the bytes for the MCU's of the next restart interval.

Again, performance-neutral.