LJpegDecompressor::decodeN(): do proper restart marker checking
We shouldn't really blindly hope that those two bytes
are what we hope they are, we should verify that they are a restart marker,
with correct number (modulo 8).

I'm not sure if there are cases where we don't exhaust the entirety
of the input buffer beforehand, as in, may there be padding bytes
before said restart marker?

Again, performance-neutral for the normal case.