oss-fuzz: build our own libc++-16
oss-fuzz comes with LLVM15-ish, but that (aside from macOS,
which comes with LLVM16), is the only libc++-based distro,
so it looks like we can get away with requiring libc++-16.

It is a pity that we have to support clang-15,
but that is much more invasive of a change.

Refs. https://github.com/google/oss-fuzz/issues/9989